Add files via upload
The aim of this project is to develop a real-time speech-to-text multilingual converter that helps users work across languages very effortlessly. It captures real-time spoken input, and before recording, it asks the user to choose the input language. Then, such speech is translated into the target output language, giving both a text and optional audio output of the translation. The user can choose any of the three text-to-speech machines, which are for an audio experience-pyttsx3 and gTTS. Volume and words-per-minute settings are also provided for fine-tuning the output to improve clarity and comfort. This project is an actual presentation of advanced deep learning models in addressing precise speech recognition and translation. The system holds the promise of enhancing multilingual interaction in customer support, education, and international collaboration, among other domains.